+++ test/unit/mail_2_rmq/process_file.py
@@ -475,7 +475,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mail_2_rmq.archive_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
     @mock.patch("mail_2_rmq.rabbitmq_class.create_rmqpub")
-    def test_false_false_connect2(                      # pylint:disable=R0913
+    def test_false_false_connect2(              # pylint:disable=R0913,R0917
             self, mock_rmq, mock_log, mock_archive, mock_rm, mock_attch):
 
         """Function:  test_false_false_connect2
@@ -504,7 +504,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mail_2_rmq.archive_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
     @mock.patch("mail_2_rmq.rabbitmq_class.create_rmqpub")
-    def test_false_false_connect(                       # pylint:disable=R0913
+    def test_false_false_connect(               # pylint:disable=R0913,R0917
             self, mock_rmq, mock_log, mock_archive, mock_rm, mock_attch):
 
         """Function:  test_false_false_connect
@@ -559,7 +559,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mail_2_rmq.archive_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
     @mock.patch("mail_2_rmq.rabbitmq_class.create_rmqpub")
-    def test_false_true_connect2(                       # pylint:disable=R0913
+    def test_false_true_connect2(               # pylint:disable=R0913,R0917
             self, mock_rmq, mock_log, mock_archive, mock_rm, mock_attch):
 
         """Function:  test_false_true_connect2
@@ -587,7 +587,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mail_2_rmq.archive_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
     @mock.patch("mail_2_rmq.rabbitmq_class.create_rmqpub")
-    def test_false_true_connect(                        # pylint:disable=R0913
+    def test_false_true_connect(                # pylint:disable=R0913,R0917
             self, mock_rmq, mock_log, mock_archive, mock_rm, mock_attch):
 
         """Function:  test_false_true_connect
@@ -641,7 +641,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mail_2_rmq.archive_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
     @mock.patch("mail_2_rmq.rabbitmq_class.create_rmqpub")
-    def test_true_false_connect2(                       # pylint:disable=R0913
+    def test_true_false_connect2(               # pylint:disable=R0913,R0917
             self, mock_rmq, mock_log, mock_archive, mock_rm, mock_attch):
 
         """Function:  test_true_false_connect2
@@ -669,7 +669,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mail_2_rmq.archive_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
     @mock.patch("mail_2_rmq.rabbitmq_class.create_rmqpub")
-    def test_true_false_connect(                        # pylint:disable=R0913
+    def test_true_false_connect(                # pylint:disable=R0913,R0917
             self, mock_rmq, mock_log, mock_archive, mock_rm, mock_attch):
 
         """Function:  test_true_false_connect