+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [2.1.0] - 2025-06-24
+- Added additional checks when the attachment is a text file.
 - Added process ID to all log entries to be able to track each run of the program.
 - Updated python-lib to v4.0.2