+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [2.2.0] - 2025-09-05
+- Updated python-lib to v4.0.3
+- Updated rabbitmq-lib v2.4.1
+
+### Changes
+- Documnetation changes.
+
+### Deprecated
+- Support for RabbitMQ v3.6
+
+
 ## [2.1.0] - 2025-06-24
 - Added additional checks when the attachment is a text file.
 - Added process ID to all log entries to be able to track each run of the program.